make mcp.TranslatorPlugin type public
