+++ go/pkg/mcp_translator/mcp_translator_types.go
@@ -1,0 +1,14 @@
+package mcp_translator
+
+import (
+	"context"
+
+	"github.com/kagent-dev/kagent/go/api/v1alpha1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+type TranslatorPlugin func(
+	ctx context.Context,
+	server *v1alpha1.MCPServer,
+	objects []client.Object,
+) ([]client.Object, error)
